cmd.fnl -> godbolt.fnl
